Add page refresh when post is deleted.

+++ devblog/devblog/ClientApp/src/components/DeletePost.tsx
@@ -1,24 +1,26 @@
 import IPost from "../interfaces/IPostProps";
 import { MdDelete as Trash } from "react-icons/md";
 import { GetIsAdmin } from "../components/AuthenticationService";
-import { useNavigate } from "react-router-dom";
 import "./styles/DeletePost.css";
 import { useEffect, useState } from "react";
 
 const DeletePost = (props: IPost) => {
     const [isAdmin, setIsAdmin] = useState(false);
-    const navigate = useNavigate();
 
     const handleDelete = () => {
-        fetch(`api/posts/${props.id}`, {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json",
-                "Authorization": `Bearer ${localStorage.getItem("token")}`
-            }
-        }).then(() => {
-            navigate("/");
-        });
+        const shouldDelete = window.confirm("Are you sure?");
+
+        if (shouldDelete) {
+            fetch(`api/posts/${props.id}`, {
+                method: "DELETE",
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${localStorage.getItem("token")}`
+                }
+            }).then(() => {
+                window.location.reload();
+            });
+        }
     };
 
     useEffect(() => setIsAdmin(GetIsAdmin), []);
